fix: 답글, 댓글 엔티티 통합
#3
