+++ src/main/java/com/koreanbrains/onlinebutlerback/entity/comment/Comment.java
@@ -11,6 +11,7 @@ import lombok.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
+@Table(indexes = @Index(name = "idx_parent_id", columnList = "parent_id"))
 public class Comment extends BaseTimeEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -25,6 +26,10 @@ public class Comment extends BaseTimeEntity {
     private String text;
 
     @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
     private Comment parent;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Comment root;
 
 }
