+++ src/main/java/com/koreanbrains/onlinebutlerback/repository/comment/CommentRepository.java
@@ -9,4 +9,6 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
 
     void deleteByPost(Post post);
 
+    void deleteByRoot(Comment comment);
+
 }
