--- src/main/java/com/koreanbrains/onlinebutlerback/repository/comment/ReplyCommentRepository.java
@@ -1,10 +1,0 @@
-package com.koreanbrains.onlinebutlerback.repository.comment;
-
-import com.koreanbrains.onlinebutlerback.entity.comment.Comment;
-import com.koreanbrains.onlinebutlerback.entity.comment.ReplyComment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReplyCommentRepository extends JpaRepository<ReplyComment, Long> {
-
-    void deleteByComment(Comment comment);
-}
